% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorDeviationAll.R
\name{direction_ac}
\alias{direction_ac}
\title{Direction Allocation Component}
\usage{
direction_ac(vdata)
}
\arguments{
\item{vdata}{a numeric data frame of four columns; in the format: x magnitude, x direction, y magnitude, y direction.}
}
\description{
Computes the allocation component of direction deviation between two vector data sets.
Allocation component is calculated as the Mean Absolute Deviation minus the Quantity Component.
}
\details{
Equations and method for this function are based on Robert Pontius' vector deviation equations in \emph{Metrics that make a difference.}"
}
