% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorDeviationAll.R
\name{direction_md}
\alias{direction_md}
\title{Direction Mean Deviation}
\usage{
direction_md(vdata)
}
\arguments{
\item{vdata}{a numeric data frame of four columns; in the format: x magnitude, x direction, y magnitude, y direction.}
}
\description{
Computes the mean deviation of direction between two vector data sets.
If X or Y avg direction is NA, returns zero.
}
\details{
Equations and method for this function are based on Robert Pontius' vector deviation equations in emph{Metrics that make a difference.}"
}
