% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorDeviationAll.R
\name{input_uv}
\alias{input_uv}
\title{Convert Table of u and v Components into Vector Components}
\usage{
input_uv(data)
}
\arguments{
\item{data}{table of u and v component of winds from 2 variables; must be in form variable 1 u, variable 1 v,
variable 2 u, variable 2 v.}
}
\description{
Converts a table of u and v component winds of 2 variables into vector
components. Input table must be in form variable 1 u, variable 1 v,
variable 2 u, variable 2 v.
Prepares meteorological wind data from raster files for use in vector deviation calculations.
}
\details{
Conversion of u and v component wind into vectors maintains and combines directions into
the direction attribute, and speed into the magnitude attribute.
}
