% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorDeviationAll.R
\name{mad_plot}
\alias{mad_plot}
\title{Mean Absolute Deviation Plot}
\usage{
mad_plot(vdata)
}
\arguments{
\item{vdata}{a numeric data frame of four columns; in the format: x magnitude, x direction, y magnitude, y direction.}
}
\description{
Creates a stacked bar graph of quantity and allocation components of MAD
for both magnitude and direction of two vector variable sets.
}
\details{
Equations and method for this function are based on Robert Pontius' vector deviation equations in \emph{Metrics that make a difference.}"
}
