% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorDeviationAll.R
\name{magnitude_mad}
\alias{magnitude_mad}
\title{Mean Absolute Deviation of Magnitude}
\usage{
magnitude_mad(vdata)
}
\arguments{
\item{vdata}{a numeric data frame of four columns; in the format: x magnitude, x direction, y magnitude, y direction.}
}
\description{
Computes the mean absolute deviation (MAD) of magnitude between two vector data sets.
Mean Absolute Deviation is calculated as the sum of Allocation and Quantity Components.
}
\details{
Equations and method for this function are based on Robert Pontius' vector deviation equations in \emph{Metrics that make a difference.}"
}
