% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorDeviationAll.R
\name{magnitude_md}
\alias{magnitude_md}
\title{Magnitude Mean Deviation}
\usage{
magnitude_md(vdata)
}
\arguments{
\item{vdata}{a numeric data frame of four columns; in the format: x magnitude, x direction, y magnitude, y direction.}
}
\description{
Computes the mean deviation of magnitude between two vector data sets.
Uses the equation mean vector of \code{y} - mean vector of \code{x}.
}
\details{
Equations and method for this function are based on Robert Pontius' vector deviation equations in \emph{Metrics that make a difference.}"
}
