% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorDeviationAll.R
\name{magnitude_qc}
\alias{magnitude_qc}
\title{Magnitude Quantity Component}
\usage{
magnitude_qc(vdata)
}
\arguments{
\item{vdata}{a numeric data frame of four columns; in the format: x magnitude, x direction, y magnitude, y direction.}
}
\description{
Computes the quantity component of magnitude deviation between two vector data sets.
Quantity component is calculated as the absolute value of magnitude mean deviation.
}
\details{
Equations and method for this function are based on Robert Pontius' vector deviation equations in \emph{Metrics that make a difference.}"
}
