% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorDeviationAll.R
\name{raster_input_uv}
\alias{raster_input_uv}
\title{Raster to Table Input (u/v component)}
\usage{
raster_input_uv(m1u,m1v,m2u,m2v)
}
\arguments{
\item{m1u}{raster of u component of variable 1}

\item{m1v}{raster of v component of variable 1}

\item{m2u}{raster of u component of variable 2}

\item{m2v}{raster of v component of variable 2}
}
\description{
Converts 4 input maps of u and v components of wind into a table of vector components (direction and magnitude).
Prepares meteorological wind data from raster files for use in vector deviation calculations.
}
\details{
Accepts any raster map file type supported by the raster package.
Conversion of u and v comonpent wind into vectors maintains and combines directions into
the direction attribute, and speed into the magnitude attribute.
}
