% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VectorDeviationAll.R
\name{raster_input_vector}
\alias{raster_input_vector}
\title{Raster to Table Input (vector)}
\usage{
raster_input_vector(m1,d1,m2,d2)
}
\arguments{
\item{m1}{raster of magnitude component of vector variable 1}

\item{d1}{raster of direction component of vector variable 1}

\item{m2}{raster of magnitude component of vector variable 2}

\item{d2}{raster of direction component of vector variable 2}
}
\description{
Converts 4 raster files of 2 vector variable sets into table for use in deviation
calculations. Expects a magnitude and direction component raster for each vector set.
}
\details{
Accepts any raster map file type supported by the raster package. Rasters must be
input in the order specified. Prepares data to be used in vector deviation functions.
}
